Fixed create community page
create_community.dart no longer a file, widget for it can be found in create_community_page.dart
