+++ lib/constants/routes.dart
@@ -1,8 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:linear/auth_pages/sign_in.dart';
 import 'package:linear/auth_pages/sign_up.dart';
-// import 'package:linear/pages/community_page/community_page.dart';
-import 'package:linear/pages/create_community_page/create_community_page.dart';
+import 'package:linear/pages/common_widgets/create_community_page.dart';
 import 'package:linear/pages/home_page.dart';
 import 'package:linear/pages/profile_page/profile_page.dart';
 import 'package:linear/pages/search_page.dart';
@@ -15,6 +14,5 @@ Map<String, Widget Function(BuildContext context)> linearRoutes = {
   '/search': (context) => const SearchPage(),
     '/goals': (context) => const GoalPage(),
   '/profile': (context) => const ProfilePage(),
-  // '/community': (context) => const CommunityPage(),
   '/createCommunity': (context) => const CreateCommunityPage(),
 };
